Adds USCM Patch to Uniforms (#5868)
# About the pull request
Adds the USCM ~and Falling Falcons~ patch automatically to vended
uniforms.

**EDIT: Per the Technical Manual as @morrowwolf shared, only the USCM
patch is added now.**

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Get on the ready line Marines!

This is a straightforward PR imo, it looks cool and is lore accurate.

Like rank pins, the ~unit patch~ and USCM patch are a standard part of
the USCM uniform and all Marines aboard the Almayer would be wearing
one. This can be found in the movies as well as technical manuals for
the Marines. It also adds some “esprit de corp” for the Falling Falcons
USCM, now every player down to the newest one can take pride fighting
for liberty.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/32654903/f87847f8-12a1-4aff-85e0-d423b2669e61)

</details>


# Changelog

:cl:
NervanCatos
add: Uniforms vend with USCM patch.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>