Body zone targeting keybinds refactor and a fix (#5943)
# About the pull request

Refactors the bodypart selection keybinds to be a bit more
"object-oriented", and to fix a bug when it tries to find the UI
element.

# Explain why it's good for the game

Makes the code very slightly cleaner, and a bit less prone to error.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
refactor: Refactored the bodypart selection keybinds.
fix: Fixed a bug where selecting a new body zone would sometimes modify
the wrong HUD element.
/:cl: