Minor additons to Synth vendor, renames a uniform used by colonists (#6010)
# About the pull request

This PR adds 2 items to Synth Vendor, a purple armband and a requested
white suit & tie.



This suit & tie was previously named "Captain's Service Uniform" and
used for years as a colonist uniform - specifically the plain clothes WY
Security Guard. The name is weird.

However, at some point, it was renamed to Commanding Officer's Service
Uniform and this makes the name even worse. Plain clothed WYSec are the
guys in suits next to CLs who don't break the mood with their ugly
orange or white more police/militarized counterparts. To be honest i
have attended a lot of CO events and I've never seen anyone wear this
LOL
It is a suit and tie with black slacks, so I understand why it was added
originally. The distinction of it being a dress down uniform is
unnecessary

I added a counterpart so we don't have to have this distinction while
keeping the classic colonists in.

# Explain why it's good for the game

This separates the colony/synths from COs so that they don't have to
share names, even if its rarely used by the latter. More immersive this
way

# Testing Photographs and Procedure
[<!-- Include any screenshots/videos/debugging steps of the modified
code functioning successfully, ideally including edge cases. -->
<details>
[<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

https://cdn.discordapp.com/attachments/975247220960014386/1221646023261294632/image.png?ex=66135585&is=6600e085&hm=c1eb0ac96f2f9feb992d3f8fbe9c4976dc2854be16ecdf4f6aee708cb7bb90cc&
</details>


# Changelog
:cl:
add: adds a purple armband and uniform to Synth Vendor
add: adds a renamed uniform commonly used by colonists with a better
name
/:cl: