Fix,update CLF ERT Station. fix space tile and missing platform. (#5992)
# About the pull request
Fix having space bellow the door of the temple.(their was some weirdness
when the wall next to it when they got blow up their was space bellow
not sure what cause it...)
added missing platform
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: replace space bellow door in the temple by proper floor.
maptweak: added a missing platform part.
/:cl:

---------

Co-authored-by: Julien <jverger.ingx@gmail.com>