Fixes carlarc's OP chocolate (#6003)
# About the pull request
Fixes #4695, The issue isn't inherently the attack speed, although it is
quite fast, many small items also have a relatively fast attack speed.
The problem is that when you combine the attack speed with a food item
that somehow has a force of 35 you get a very effective weapon capable
of destroying many different types of structures fairly easily, same
issue with the scalpel.

My solution here was to just decrease the amount of force the items have
rather than lowering the attack speed. You shouldn't be able to use a
scalpel or a box of chocolate to more effectively destroy weeds rather
than a machete. I doubt these are the only two items that have this
problem, I guess in the future when someone comes across a similar issue
just link this pr so they can easily identify the exploit.

# Changelog

:cl: 
fix: lowers the scalpel and chunk box's ability to destroy various
structures.
add: chunk box is now a melee weapon
/:cl:

---------

Co-authored-by: Vicacrov <49321394+Vicacrov@users.noreply.github.com>
Co-authored-by: SabreML <57483089+SabreML@users.noreply.github.com>
Co-authored-by: DOOM <N/A>