Project ARES: Added Tech Unlock logs (#5006)
# About the pull request
As the title says, you can now view who unlocked techs in the ARES
Interface (if you're senior command)
I have also split the cameras in the core to an indivual network that is
not shared with the rest of the ship, and placed a camera console to
view this network in CIC, CO's Office, CE's Office and Brig Office
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added AI Core camera consoles to CE, CO and Brig offices, along
with CIC and Lifeboat bubble.
maptweak: Added Tech Control Console to ARES Core.
add: Added Tech Unlock logs to ARES Interface
/:cl: