Fix inconsistent number ranges for EFTPOS pin & access code (#6050)
# About the pull request

This PR fixes a couple places where tgui_input_number is called with
default arguments when it should be within a specific range for the
EFTPOS when specifying the access code and pin.

# Explain why it's good for the game

Certain values can currently lock you out of using the EFTPOS.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
fix: Fixed some number entry for your pin and access code in the EFTPOS
/:cl: