Makes acid spray extinguish more consistent (#6053)
# About the pull request
Acid spray used to only extinguish you if you were on the same tile when
it is initialized rather than when crossing it as well, this PR makes it
so it extinguishes you when crossing it as well.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Extremely weird and unintuitive when you walk over 3 tiles of acid spray
and don't get extinguished.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: Acid sprays will now extinguish fire from other xenomorphs more
consistently
/:cl: