[Balance]add the ability to store fixOvein in the pouch medkit. (#6037)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
fix ovein is an upgrade version of the suture to fix IB so it make sense
to me to allow medic to store it with the suture.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: add the ability to store fixOvein in the pouch medkit.
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>