reduces impact area from 9x9 to 7x7, lowers DPS slightly (#6031)
# About the pull request

okey this is basicly copy of my suggested buff from more then a year
ago, but now it is a nerf. Main complain on why the GAU is so OP and
strong is its spreat ( that used to be the reason why it was so weak and
shit). Therfore, lets do some math. Current situation is that the GAU
fires 80 rounds per burst at area of 81 tiles( 9x9) ~0,988 round per
tile, the proposed change aims the keep the damage on anyone standing in
the impact area about the same ( it is a bit lower might need
compensation with bringing the damage to pre double bullet values) , by
bringing down the impact area to 49 tiles (7x7) and lowering the number
of shots fired to 40, we get value of ~0,816 round per tile. meanwhile
it keeps the rest like how long the bursts last the just as long.
(https://github.com/cmss13-devs/cmss13/pull/4302 the old "buff")

# Explain why it's good for the game
Making two bullets per bullet was not a bad idea in general, it brought
the DPS on competetive level with other weapons, but the huge spread
leads to too big compared to other ordinance. Bringing it in line with
other ordinance should allow for easier CAS balance around smaller
impact FMs.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: reduced GAU spread from 9x9 to 7x7
balance: reduced GAU shots fired per shot back to 1 (bullets per tile
from 0.99 to 0.82)
/:cl:

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>