No More Prep Raiding (#6083)
# About the pull request

- Removes access to Marine Prep and Marine Prep Ammo vendors for Cargo
Technicians (CT).
- Doubles regular ammo in Requisitions to compensate changes and
eliminate the need for Cargo Techs to pilfer Marine Prep of all its
ammo.
- Quartermaster (QM) is unchanged.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

**I would consider this a HOTFIX**, Maintainers update the title as you
see fit, I just didn't want to be _that guy_.

### PRIORITY ISSUE: 
By mid-round of medium pop or higher the ColMarTech Automated Weapons
Rack is void of almost all ammo. This is a priority to fix because it
_ruins_ the new player experience as they have no ammo to use for their
primary ammo before deploying. This causes frustration and can/has led
to new players quitting CM before they even got to start. The root
causes of this is are:
A. CTs raiding the Marine Preps after first drop for more ammo.
B. The round scale not reflecting the current population changes and/or
ammo expenditures.

To fix this CT access has been removed from the Weapons Rack to
eliminate the possibility of CTs pilfering the ammo from new players. To
compensate, Regular Ammo in Requisitions has been _doubled_ there should
no longer be a need to raid Marine Prep for ammo. The QM still has
access to the vendor for edge cases where access to the Weapons Rack is
needed, but they should not abuse this retention of access.

### SECONDARY ISSUE:
There should be no reason for CTs to have to raid the Marine Preps for
ammo in the first place, logically this makes no sense. Requisitions is
the nexus of all weapons and ammo on the Almayer. For **Regular Ammo**
the logistical difficulty should be _getting_ the ammo to the frontlines
for resupply, not artificial scarcity of the ammo.

This PR will ideally fix this issue as the round start regular ammo
count has been doubled. **This value is subject to change.**
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>
I tested it, they don't have access and the round scale is updated to
the new values.
</details>


# Changelog
:cl: NervanCatos
balance: removes CT access to marine prep weapons racks, doubles ammo
count of all regular ammo in req to compensate.
/:cl: