Cryo Tube Rework (#6058)
# About the pull request

This PR reworks the cryotubes. The effect of diluting or multiplying
chemicals is no longer possible, but now the use of cryoxadone or
clonexadone in a cryotube can allow chemicals to process in the dead.
Ultimately usage of the cryo chemicals will be much higher, and you need
to be careful not to introduce too many chemicals into your patient
because the dosage is 5u whenever cryoxadone or clonexadone is available
and can be administered.

Cryotubes also now actually require power, update their icon if they
lose power, and use the two-ping sfx for warnings (such as now warnings
as the patient's revive timer is running out). For the critical
messages, it will make an alert if the patient is dead, then once they
hit the 50% mark (2.5 mins), then at the 1 minute mark.

See testing for a demonstration.

# Explain why it's good for the game

Cryotubes in general are very underused with the exception of exploiting
how it can multiply chemicals. This should give them a unique feature to
medbay.

It is likely that this may prove to be too powerful, but it puts more
demand on chemistry to provide the cryoxadone or clonexadone in order to
do so.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Initial testing:
https://youtu.be/PA43YvzDRVM

Critical (dead state) alert testing:
https://youtu.be/PLjQxzv28l8

Notification changes:

![release](https://github.com/cmss13-devs/cmss13/assets/76988376/325e852a-79ee-4a68-9c3c-cd9001f6cc0f)

Power fixes:

![power](https://github.com/cmss13-devs/cmss13/assets/76988376/afd952eb-3004-4c91-b255-83aa98fd539f)

</details>


# Changelog
:cl: Drathek
balance: Cryotubes no longer multiply chemicals.
balance: Cryotubes can now allow chemicals to process in dead if
cryoxadone or clonexadone is present.
fix: Fixed cryotubes not actually updating their icons/power usage when
they become unpowered.
fix: Fixed cryotubes healing despite inoperable.
fix: The cryotube Eject Occupant verb now works for people other than
the occupant (useful if inoperable)
ui: Cryotubes can optionally announce again on the medical channel.
/:cl: