Adds an APOLLO maintenance controller to the Warden's Office. (#6043)
# About the pull request

Simply adds an APOLLO maintenance controller to the MW office in the
brig. Moves some of the buttons down a bit to accommodate new computer
console. Screenshot provided in the relevant section.

# Explain why it's good for the game

The brig, despite being a zone that's prone to all sorts of maintenance
issues, has no APOLLO maintenance controllers at all. The nearest one
being at the CIC. This should give both wardens and WJs something to do
when it's quiet.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/42235601/58267043-1313-48a1-934e-429430704604)

</details>


# Changelog
:cl:
mapadd: Added an APOLLO maintenance controller to the Warden's Office.
/:cl: