Xenothumbs Minor Event Button Comeback (#6034)
# About the pull request
Working version that is TMable for april fools. 


Revived and fixed up
https://gitlab.com/cmdevs/colonial-warfare/-/merge_requests/1915/diffs#9ec310cca070259d82839a4a70e89dd858ded438_958_958

https://github.com/cmss13-devs/cmss13/pull/2997

Additions from  Vanagandr's original MR
- xenos can use health items on marines
- vendors are usable now, all checks are simply skipped so introducing
xenos with thumbs to snowflake vendors is a horrible idea (as if
enabling this in the first place wasn't bad enough) [needs further
testing]
- xenos can drive
- M2C usable


I think this is pretty much done, it's already exceeding the scope of
the original and we're just making xenos human past this point

Original description
> TL;DR: https://streamable.com/rlvs86
> 
> Mainly this allows xenos to pick up items. They can then use most of
them.
> 
> Ex. a xeno can't flip a plasteel cade open - but if she has a toolbelt
and time to work, she can unwrench it.
> 
> They can fumble about but a lot of QoL and objects (middleclick
container open, click-dragging containers, tables, etc.) were never
designed to be interacted with by xenos so item interactions are a bit
clunky.
> 
> They can't strip clothing and gear from people.
> 
> They *can* strip and attach attachments but can't aim down scopes.
> 
> They *can* FF.
> 
> They *can't* use MGs.
> 
> IFF works, roughly, not picking up on specific hives but allowing them
to shoot past other xenos and hit humans.
> 
> They pass spec checks and can clumsily handle medic gear.
> 
> needless to say this is all colossally imbalanced and has potential
for fuckery limited only by the xeno's imagination and whether the thing
they're messing with makes an explicit ishuman() check. Or is a
structure. They can't hand-interact with structures, so they can shove
shells into a mortar but can't change its coords.
> 
> Cosmetically, gun inhands don't line up with xenos. Tweaked it a bit
to center it horizontally but lining inhands up with hands for every
xeno and xenonid sprite is a bit more hassle than seems worthwhile ATM.
> 
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
TM funny
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:  Vanagandr, totalepicness
add: Added a button to event panel to give xenomorphs opposable thumbs
and firearms permits, to individuals, to hives, or to all xenos
everywhere. They're clumsy, though, and have issues with fine
manipulation sometimes. Remember that warriors cannot throw objects and
make for poor grenadiers.
add: Xenos with opposable thumbs can use buckles, drive vehicles, use
machine guns and vendors without care for access or skills
/:cl:

---------

Co-authored-by: Epicness <coolguyethanw@gmail.com>
Co-authored-by: harry <me@harryob.live>