Requisition Centralization [Part 1.5] AKA medvendor infinite restock outside medbay cheese removal redux (#5677)
# About the pull request

This PR adds a new structure: the medical supply link port. This is a
structure that allows medical vendors to link to the Almayer's internal
medical supplies and have larger restocking capabilities. They can be
found around the Almayer where existing medical vendors are.

This PR makes it impossible to restock/refill non full stacks at medical
vendors which are not linked to a medical supply link port

This PR makes it impossible to restock/refill non full autoinjectors and
bottles at medical vendors which are not linked to a medical supply link
port

revives #2155

# Explain why it's good for the game

This PR is good for the game because it will get req more involved in
the medical supply process, as well as limiting the capability of medics
to instantly resupply from a vendor at the fob or in hydro. This makes
it so that medics have to more carefully pick their supplies, and plan
ahead for resupplying, as well as consider other medics' use of
supplies, as nothing is infinite groundside anymore.

Overall, it will make the medic supply gameplay loop more engaging and
interesting, as well as reducing marine capability to infinitely heal
and resupply themselves groundside.

Makes #5490 more useable

# Testing Photographs and Procedure

<details>

https://gyazo.com/d9e8e3b63daab1241e046c23bb5dd140

</details>


# Changelog

🆑 Stan_albatross,kugamo,LTNTS,Drathek
balance: medical vendors, excepting nanomeds, can now only
refill/restock stacks and autoinjectors/bottles when on special medical
supply link ports that can only be found shipside. Sprite by kugamo.
fix: supply pads will no longer generate ambient occlusion.
/:cl:

---------

Co-authored-by: Nanu308 <59782240+Nanu308@users.noreply.github.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>