Auto Delete Foodless MREs (#5761)
# About the pull request

MRE's that contain no food will now be automatically deleted.

# Explain why it's good for the game

Current MRE's mean the FOB, and frontline where a resupply crate lands,
are littered with endless amounts of opened MRE's that are filled with
the fluff items. This not only clutters up the area but it means if you
are desperate for food, you have to search these half-used MRE's.

With this change, this means that a MRE on the ground will actually
contain food instead of matches or water, and this will significantly
declutter the FOB and frontline of food items.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: MRE's will now auto-delete if they do not contain and food items
when they are discarded.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>