'Search by key' button for the Keybind menu (#6070)
# About the pull request

Adds a 'search by key' button to the Keybind menu, which scrolls the
window to the first entry bound to the inputted key.

Ideally this would somehow loop through everything that the key is bound
to, but I haven't been able to figure that out yet.

# Explain why it's good for the game

It's easy to search for the name of a keybind if you already know what
you're looking for, but if for example you're trying to figure out what
the 'F' key is doing, or whether or not a key you want to bind is
already in use, then currently you need to look through the whole list
manually.

# Testing Photographs and Procedure
<details open>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/57483089/791a6855-1b23-40fd-892b-873032956ec4

</details>


# Changelog
:cl:
ui: Added a 'search by key' button to the Keybind menu.
/:cl: