Buffs Dropship Crews, required one medical hour (#6093)
# About the pull request

(cleaned up ver of #6092 + using better code from ihatethisengine)
This PR adds in a medical hour requirement for dropship crew chiefs, as
they have medical training and should are required to be knowledgeable
on medical.
It also adds in the ability for dropship crew to fix their doors with a
multitool. Instead of the usual five seconds for engineering trained
personnel, dropship crew take eight seconds.
Finally, a tip is added,

> "The Queen can unbolt and break dropship doors by prying them open,
even if they are unlocked. Pilots, Dropship Crew Chiefs, Engineers and
Synths can repair these doors with a multitool."

# Explain why it's good for the game

Since Dropship Crew Chiefs have surgery knowledge, they must have a
buffer to make sure that people who play Dropship Crew Chief are
knowledgeable on medicine. This is the same for medics and doctors.

Next, it seems kinda stupid that Dropship crew aren't able to fix the
own doors to their own dropship, and adds to the stigma that transport
pilots and DCC's are un-fun to play, ignored and useless. By making it
so they can actually repair their own dropships, they can bring a little
more usefulness to rounds.

Finally, a tip is added to accommodate this PR, as I'm pretty sure no
one even knew you CAN repair queen destroyed doors. I'd like to add a
description to broken doors, but I'm not sure if that's possible with
doorcode (im not very smart).

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

it works i tested it i think

</details>

# Changelog

:cl:
add: Adds a one hour medical timelock for DCC's
add: Adds a tip of the round to accommodate for this PR and #5808
balance: DCCs and Pilots can now fix queen broken dropship doors
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>