flames remove snow !  (#6001)
# About the pull request

makes snow scortchable and implements behavior for it ( thank you kugamo
for making the LV grass scortchable so that the basework for it exists).
intensity numbers might need tweeking if maintainers feel like it, but
other than that it works.

# Explain why it's good for the game

removing snow is pain in ass and even more so for marines, explosions
can already remove snow and it only makes sance for flames to melt the
snow. only bluefire can remove the deapest snow in one use, green and
red flames remove abut two layers


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: flames melt the snow 
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>
Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>