Preferences for a darker crit overlay and a black flash overlay (#5997)
# About the pull request

Now these overlays have less of a shock factor for accessibility. The
reasoning for also a preference for crit is because when being revived
you might still be deep in crit which can also be a flashbang. These
could be combined into one preference, but I feel it didn't hurt to keep
them separate so they're more obvious what they affect.

- For dark crit preference, it is the original overlay but colored gray
so lessen its brightness.
- For dark flash preference, the screen will turn black instead of
white.

See photographs below.

# Explain why it's good for the game
It was requested that there be a preference to lessen the flash effect
for people that are photosensitive.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/6bec8f49-4153-4312-a149-416364982d88)


![normal](https://github.com/cmss13-devs/cmss13/assets/76988376/cad1339c-9e6f-45fb-9b73-89cac3d41686)

![dark](https://github.com/cmss13-devs/cmss13/assets/76988376/ed968a81-cbe0-4f1e-ad50-7a26735b5d7d)


![flash_normal](https://github.com/cmss13-devs/cmss13/assets/76988376/cfdfb6de-e251-4d8a-81a7-ae164c26fa18)

![flash_dark](https://github.com/cmss13-devs/cmss13/assets/76988376/fd22dc72-68b9-47dd-a6f1-4fcb249a669d)

</details>


# Changelog
:cl: Drathek
add: Added preference for a darker crit overlay and black flash overlay
/:cl:

---------

Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>