Small fix to Nightmare documentation example (#6072)
# About the pull request
Fixes the documentation regarding the nightmare landmarks example.
Viro_open which is presented as the example, fails to inform you that
the text after the _ will not be used by the landmark and is simply
additional info for the viewer. This has now been rectified with this
PR.


# Changelog
:cl:
fix: Fixes Nightmare documentation example.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>