Fix additional cryo reagent duplication via TOUCH reaction (#6122)
# About the pull request

This PR is a follow up to #6058 where it was brought to my attention
that chemicals were still duplicating from cryotube beakers. This is
because the cryotube was trying to simulate the patient being immersed
in the reagents, but with a 50% permeability that doesn't even consume
the reacted reagents, it easily gets out of control if large volumes are
involved and or the ratio of cryo meds is small in the mixture causing
frequent reapplications of reagents.

# Explain why it's good for the game

It was intended to remove chemical duplication via cyropods. This shores
up another source of chemical duplication.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![cryo](https://github.com/cmss13-devs/cmss13/assets/76988376/26b19859-9d86-48e4-8800-be34089cf1d8)

After rework 200 bicaradine and 20 cryoxadone:


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/3551f146-f874-4700-bf67-c7a9f2aebb87)

</details>


# Changelog
:cl: Drathek
fix: Fixed cryotubes still duplicating chemicals because of touch
reactions
/:cl: