Better throw logic (#6040)
# About the pull request
Throw will now attempt to reach the target rather than the a predefined
path that could change if thrown towards an object

Tested on local but this affect a lot of stuff so needs to be tested
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
When you try to throw something and the target slightly moves one
nanosecond after you throw it sucks
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Throws will now attempt to reach their target rather than the turf
they were on when thrown
/:cl: