Adds the Quebec flag (#6133)
# About the pull request

Adds the Quebec flag to the game, similar to the Scotland one.
This one actually is apparently fuckall nowhere so I had to find the
original source of the flags, download it, and make it myself -
https://github.com/gosquared/flags
And even then, it didn't turn out perfect, the fleur de lis' weren't all
showing up. I had to sorta blow them up for them to appear nice in OOC.

# Explain why it's good for the game

Basically the same reason as
https://github.com/cmss13-devs/cmss13/pull/640
I do not consider myself to be a Canadian and would rather be able to
use my own country flag. This makes the game playable (for me)!

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/53777086/7fcb386b-070a-4a50-a933-2a14ffe31dcb)

</details>


# Changelog

:cl:
imageadd: Added the Quebec flag to OOC for all Quebecois.
/:cl: