Disable un-resting someone when they are dizzy (#6135)
# About the pull request

This PR prevents unresting someone when you pat them if they are dizzy.

# Explain why it's good for the game

Voluntary resting disables screen shifting and allows dizziness to pass
faster. Other players won't necessarily know if the rest is voluntary,
but this change makes it so they get all the effects of patting without
the unresting aspect if they are dizzy.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/76988376/d0219430-8a94-4eb6-98a7-46d8abe7c45c

</details>


# Changelog
:cl: Drathek
add: Patting someone that is dizzy will no longer unrest them
/:cl: