Fix unpacking vendors not performing area entry (#6105)
# About the pull request

This PR fixes unboxing vendors not entering the area to try and connect
to the powernet. Note - this will mean all items getting picked up and
dropped are now entering/exiting areas correctly. Possible there's some
obscure new behavior with this but its unlikely.

# Explain why it's good for the game

If you unbox a machine in a powered area, it should be powered.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


https://github.com/cmss13-devs/cmss13/assets/76988376/55d47379-5780-4443-b615-f8783df59aef

</details>


# Changelog
:cl: Drathek
fix: Fixed unboxed machines not connecting to the power net
/:cl: