update to specialist vendor. (#6076)
# About the pull request
everything is in the changelog.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
updating the vendor based on the other vendor
because before spe just used all it's point on is spe ammo but now he
can spend it on what he want.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: added the option to buy restricted firearm for the specialist
SU-6 and VP78 each for 15points.
balance: added the option to buy sidearm ammunition M44 (heavy and
Marksman) and SU-6 mag for 10 point each.
balance: added the option to buy sidearm ammunition M4A3 HP and AP for 5
point each.
balance: added the option to buy a whistle for 5 point and a data
detector for 10 point.
balance: added the option to buy binoculars for 5 point range finder for
10 point and laser designator for 15 point.
/:cl:

Co-authored-by: Julien <jverger.ingx@gmail.com>